refactor: se ha actualizado el nombre del dominio auth -> register para una mejor consistencia en los servicios que provee
