+++ app/modules/register/controllers/change_status.py
@@ -8,9 +8,9 @@ from uuid import uuid4
 from fastapi import HTTPException
 from sqlalchemy.orm import Session
 
-from app.modules.auth.services.email_service import EmailService
-from app.modules.auth.services.kc_service import KeycloakService
-from app.modules.auth.services.moodle_service import MoodleService
+from app.modules.register.services.email_service import EmailService
+from app.modules.register.services.kc_service import KeycloakService
+from app.modules.register.services.moodle_service import MoodleService
 from app.shared.enums.status_enum import AccountStatusEnum
 from app.shared.models.users_model import UserAccounts
 from app.shared.models.verification_tokens_model import VerificationToken
